refactor(desktop): split team hooks out of agents hooks.ts
hooks.ts crossed the 1000-line file-size ceiling (1002 lines), failing
the pre-push desktop-check gate. Move the team query/mutation hooks and
teamsQueryKey into teamHooks.ts and re-export them from hooks.ts so
existing importers are unchanged.
