+++ desktop/src/features/agents/hooks.ts
@@ -60,24 +60,16 @@ import {
   setPersonaActive,
   updatePersona,
 } from "@/shared/api/tauriPersonas";
-import {
-  createTeam,
-  deleteTeam,
-  listTeams,
-  updateTeam,
-} from "@/shared/api/tauriTeams";
+import { teamsQueryKey } from "@/features/agents/teamHooks";
 import type {
   AcpRuntime,
   AgentPersona,
-  AgentTeam,
   Channel,
   CreateManagedAgentInput,
   CreatePersonaInput,
-  CreateTeamInput,
   ManagedAgent,
   UpdateManagedAgentInput,
   UpdatePersonaInput,
-  UpdateTeamInput,
 } from "@/shared/api/types";
 import { normalizePubkey } from "@/shared/lib/pubkey";
 import type {
@@ -90,6 +82,13 @@ import type {
   ProvisionChannelManagedAgentResult,
 } from "@/features/agents/channelAgents";
 export { findReusableAgent } from "@/features/agents/agentReuse";
+export {
+  teamsQueryKey,
+  useCreateTeamMutation,
+  useDeleteTeamMutation,
+  useTeamsQuery,
+  useUpdateTeamMutation,
+} from "@/features/agents/teamHooks";
 export type {
   AttachManagedAgentToChannelInput,
   AttachManagedAgentToChannelResult,
@@ -105,7 +104,6 @@ export type {
 export const relayAgentsQueryKey = ["relay-agents"] as const;
 export const managedAgentsQueryKey = ["managed-agents"] as const;
 export const personasQueryKey = ["personas"] as const;
-export const teamsQueryKey = ["teams"] as const;
 export const acpRuntimesQueryKey = ["acp-runtimes"] as const;
 export const acpAuthMethodsQueryKey = ["acp-auth-methods"] as const;
 export const managedAgentPrereqsQueryKey = ["managed-agent-prereqs"] as const;
@@ -947,55 +945,5 @@ export function useBakedBuildEnvKeysQuery(options?: { enabled?: boolean }) {
     staleTime: Infinity,
     refetchInterval: false,
     retry: false,
-  });
-}
-
-export function useTeamsQuery() {
-  return useQuery({
-    queryKey: teamsQueryKey,
-    queryFn: listTeams,
-    staleTime: 30_000,
-    // No refetchInterval: inbound relay team changes emit `agents-data-changed`
-    // (handled by useAgentsDataRefresh). Same redundant-poll removal as
-    // usePersonasQuery.
-  });
-}
-
-export function useCreateTeamMutation() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: (input: CreateTeamInput) => createTeam(input),
-    onSuccess: (created) => {
-      queryClient.setQueryData<AgentTeam[]>(teamsQueryKey, (current) => {
-        const next = current ?? [];
-        return [created, ...next.filter((team) => team.id !== created.id)];
-      });
-    },
-    onSettled: async () => {
-      await queryClient.invalidateQueries({ queryKey: teamsQueryKey });
-    },
-  });
-}
-
-export function useUpdateTeamMutation() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: (input: UpdateTeamInput) => updateTeam(input),
-    onSettled: async () => {
-      await queryClient.invalidateQueries({ queryKey: teamsQueryKey });
-    },
-  });
-}
-
-export function useDeleteTeamMutation() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: (id: string) => deleteTeam(id),
-    onSettled: async () => {
-      await queryClient.invalidateQueries({ queryKey: teamsQueryKey });
-    },
   });
 }
